homework2/src/bg/sofia/uni/fmi/mjt/goodreads/recommender/similaritycalculator/genres/GenresOverlapSimilarityCalculator.java
package bg.sofia.uni.fmi.mjt.goodreads.recommender.similaritycalculator.genres;

import bg.sofia.uni.fmi.mjt.goodreads.book.Book;
import bg.sofia.uni.fmi.mjt.goodreads.recommender.similaritycalculator.SimilarityCalculator;

public class GenresOverlapSimilarityCalculator implements SimilarityCalculator {

    @Override
    public double calculateSimilarity(Book first, Book second) {
        if (first == null || second == null) {
            throw new IllegalArgumentException("Arguments should not be null!");
        }

        return (double) (first.genres().stream()
            .filter((String genre) -> second.genres().contains(genre)).count()) /
                Math.min(first.genres().size(), second.genres().size());
    }
    
}
